[NEW RBO] Refactoring of expressions in new RBO
